backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
mangum==0.17.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2

# CLI
rich==13.7.0
typer==0.9.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
